finalizing onboarding and organizing code
