+++ DvdScreensaver/Onboarding/DoisDedos.swift
@@ -15,7 +15,7 @@ struct DoisDedos: View {
     
     var body: some View {
         ZStack{
-            NavigationLink("",destination: ContentView().navigationBarBackButtonHidden(true) ,isActive: $shouldShow)
+            NavigationLink("",destination: DeslizeDedo().navigationBarBackButtonHidden(true) ,isActive: $shouldShow)
             
             Image("background")
                 .resizable()
